feat(threaded-replies): move replies fixtures
