+++ src/elements/content-sidebar/activity-feed/activity-feed/__tests__/ActivityThreadReplies.test.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import ActivityThreadReplies from '../ActivityThreadReplies.js';
-import replies from '../../../../../__mocks__/replies';
+import { replies } from '../fixtures';
 
 describe('src/elements/content-sidebar/activity-feed/activity-feed/ActivityThreadReplies', () => {
     const getWrapper = props => render(<ActivityThreadReplies replies={replies} isExpanded={false} {...props} />);
