+++ src/elements/content-sidebar/activity-feed/activity-feed/fixtures.ts
@@ -1,14 +1,22 @@
-import { user } from './annotations';
-
-const replies = [
+export const replies = [
     {
         created_at: '2020-09-20T00:00:00Z',
-        created_by: user,
+        created_by: {
+            id: '1',
+            login: 'johndoe',
+            name: 'John Doe',
+            type: 'user' as const,
+        },
         id: '2485934',
         message: 'First reply',
         tagged_message: 'First reply',
         modified_at: '2020-09-20T00:00:00Z',
-        modified_by: user,
+        modified_by: {
+            id: '1',
+            login: 'johndoe',
+            name: 'John Doe',
+            type: 'user' as const,
+        },
         permissions: {
             can_delete: true,
             can_edit: true,
@@ -19,12 +27,22 @@ const replies = [
     },
     {
         created_at: '2020-09-20T00:00:00Z',
-        created_by: user,
+        created_by: {
+            id: '1',
+            login: 'johndoe',
+            name: 'John Doe',
+            type: 'user' as const,
+        },
         id: '24835935',
         message: 'Last reply',
         tagged_message: 'Last reply',
         modified_at: '2020-09-20T00:00:00Z',
-        modified_by: user,
+        modified_by: {
+            id: '1',
+            login: 'johndoe',
+            name: 'John Doe',
+            type: 'user' as const,
+        },
         permissions: {
             can_delete: true,
             can_edit: true,
@@ -35,4 +53,4 @@ const replies = [
     },
 ];
 
-export default replies;
+export default { replies };
